refactors
refactoring broke some early parts; restart with these new parts
