--- engineering/sigma.c/.vscode/settings.json
@@ -1,8 +1,0 @@
-{
-    "files.associations": {
-        "stdbool.h": "c",
-        "version.h": "c",
-        "stdio.h": "c",
-        "file.h": "c"
-    }
-}
