+++ engineering/sigma.c/src/.data/sigmac.def
@@ -1,0 +1,4 @@
+<program>   := <function> ;
+<function>  := "int" <id> "(" ")" "{" <statement> "}" ;
+<statement> := "return" <expr> ";" ;
+<expr>      := <int> ;
