+++ engineering/sigma.c/src/file.c
@@ -1,19 +1,9 @@
-
-#include <stdio.h>
-#include <unistd.h>
 #include <stdlib.h>
 #include <sys/stat.h>
 
 #include "file.h"
-#include "cstring.h"
 
-#ifdef __file_h_ver
-#ifndef __file_c_ver
-#define __file_c_ver "002-a"
-#endif
-#endif
-
-size_t file_size(char *pPath)
+static size_t get_file_size(char *pPath)
 {
     struct stat buffer;
     size_t fsize = -1;
@@ -24,102 +14,14 @@ size_t file_size(char *pPath)
 
     return fsize;
 }
-const char *file_mode(file *pFile)
+
+file *file_exists(char *filepath)
 {
-    const byte mBinary = 1;
-    const byte mRead = 2;
-    const byte mWrite = 4;
-    const byte mAppend = 8;
-    const byte mCreate = 16;
-
-    const char *mode = String.empty;
-
-    //  I'm certain this can be more glamorous
-    if (pFile->mode == mRead)
-    {
-        mode = "r";
-    }
-    if (pFile->mode == mRead | mWrite)
-    {
-        mode = "rw";
-    }
-
-    return mode;
-}
-stream *file_open(file *pFile)
-{
-    if (!pFile->exists)
-    {
-        //  error
-        return NULL;
-    }
-
-    const char *mode = file_mode(pFile);
-
-    printf("open %s ['%s']\n", pFile->name, mode);
-
-    stream *pStream = malloc(sizeof(stream));
-    pStream->source = pFile->name;
-    pStream->length = pFile->size;
-    pStream->fstream = fopen(pFile->name, mode);
-    pStream->is_open = pStream->fstream != NULL;
-    pStream->pos = ftell(pStream->fstream);
-
-    // if (pFile->size > 0)
-    // {
-    //     doc->source = malloc(pFile->size);
-    //     size_t n = 0;
-    //     int c;
-    //     while ((c = fgetc(f)) != EOF)
-    //     {
-    //         doc->source[n++] = (char)c;
-    //     }
-    //     doc->source[n] = '\0';
-    // }
-
-    return pStream;
-}
-void file_close(stream *pStream)
-{
-    if (pStream->is_open)
-    {
-        fclose(pStream->fstream);
-        pStream->is_open = false;
-    }
-}
-bool file_exists(char *pPath)
-{
-    return access(pPath, F_OK) == 0 ? true : false;
-}
-file *file_new(char *pPath)
-{
-
     file *pFile = malloc(sizeof(file));
-    pFile->name = pPath;
-    pFile->exists = file_exists(pPath);
-    pFile->size = pFile->exists ? file_size(pPath) : -1;
+    pFile->name = filepath;
+    pFile->exists = access(filepath, F_OK) == 0 ? true : false;
+
+    pFile->size = pFile->exists ? get_file_size(filepath) : -1;
 
     return pFile;
 }
-//  ==================================================
-bool stream_read(stream *pStream, char *out)
-{
-    int c;
-    if ((c = fgetc(pStream->fstream)) != EOF)
-    {
-        (*out) = (char)c;
-        pStream->pos = ftell(pStream->fstream);
-    }
-
-    return c != EOF;
-}
-
-const struct File_T File = {
-    .exists = &file_exists,
-    .new = &file_new,
-    .size = &file_size,
-    .open = &file_open,
-    .close = &file_close};
-
-const struct Stream_T Stream = {
-    .read = &stream_read};
