+++ engineering/sigma.c/src/sigctypes.h
@@ -1,15 +1,11 @@
 #ifndef _SIGCTYPES_H
-#ifndef __sigctypes_h_ver
-#define __sigctypes_h_ver "002-a"
-#endif
 #define _SIGCTYPES_H
-
-#include <stdbool.h>
 
 #include <stdbool.h>
 
 #define byte unsigned char
 #define sbyte char
+#define ushort unsigned short
 
 #define glyph char
