--- sigma.c/.gitignore
@@ -1,1 +1,0 @@
-/Debug/
