--- sigma.c/.target/001/main.C
@@ -1,4 +1,0 @@
-int main()
-{
-    return 2;
-}
