--- sigma.c/.target/002/main.C
@@ -1,5 +1,0 @@
-#include io.h;
-
-int main () {
-    writef("Greeting from CTau\n");
-}
