--- sigma.c/src/sigctypes.h
@@ -1,12 +1,0 @@
-#ifndef _SIGCTYPES_H
-#define _SIGCTYPES_H
-
-#include <stdbool.h>
-
-#define byte unsigned char
-#define sbyte char
-#define ushort unsigned short
-
-#define glyph char
-
-#endif //  _SIGCTYPES_H
